backend/requirements.txt
# MetricFrame Backend Dependencies
# Production requirements file

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
pandas>=2.0.0

# Security
cryptography>=41.0.0
PyJWT>=2.8.0

# AI Providers
anthropic>=0.7.0
openai>=1.3.0
together>=0.2.0
boto3>=1.33.0
google-generativeai>=0.3.0
azure-identity>=1.15.0
